moved the .py files from resources to crosstab
